+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 All notable changes to this project will be documented in this file.
 
 
+## [0.3.0] - 21.08.2020
+### Added
+* `embedding` module with BERT pre-trained embedding model for proteins with support for single protein vectorization as well as for proteins sets which are averagend to bacteriophage level;
+
+### Changed
+* `requirements.txt` with bio-embedding and fake-useragent;
+
+
 ## [0.2.0] - 20.08.2020
 ### Added
 * `features` module with protein-based (amino-acid) feature extractor for single protein sequence as well as for multifasta;
 
 ### Changed
-* `requirements.txt` with biopython
+* `requirements.txt` with biopython;
 
 
 ## [0.1.0] - 10.08.2020